+++ Changelog.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - button for navigation to starting screen
     
+
+## [unreleased]
+
+### Added
+
+- results screen
+- widgets to display results properly
+
 ## [v0.0.3] - 2024-07-02
 
 ### Added